+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/administration/models/ModelOperationStatus.java
@@ -7,40 +7,22 @@ import com.azure.core.util.ExpandableStringEnum;
 
 import java.util.Collection;
 
-/**
- * Defines values for ModelOperationStatus.
- */
+/** Defines values for ModelOperationStatus. */
 public final class ModelOperationStatus extends ExpandableStringEnum<ModelOperationStatus> {
-
-    /**
-     * Enum value notStarted.
-     */
+    /** Static value notStarted for ModelOperationStatus. */
     public static final ModelOperationStatus NOT_STARTED = fromString("notStarted");
-    /**
-     * Enum value running.
-     */
+
+    /** Static value running for ModelOperationStatus. */
     public static final ModelOperationStatus RUNNING = fromString("running");
-    /**
-     * Enum value failed.
-     */
+
+    /** Static value failed for ModelOperationStatus. */
     public static final ModelOperationStatus FAILED = fromString("failed");
-    /**
-     * Enum value succeeded.
-     */
+
+    /** Static value succeeded for ModelOperationStatus. */
     public static final ModelOperationStatus SUCCEEDED = fromString("succeeded");
-    /**
-     * Enum value canceled.
-     */
+
+    /** Static value canceled for ModelOperationStatus. */
     public static final ModelOperationStatus CANCELED = fromString("canceled");
-
-    /**
-     * The actual serialized value for a ModelOperationStatus instance.
-     */
-    private final String value;
-
-    ModelOperationStatus(String value) {
-        this.value = value;
-    }
 
     /**
      * Creates or finds a ModelOperationStatus from its string representation.
@@ -52,11 +34,8 @@ public final class ModelOperationStatus extends ExpandableStringEnum<ModelOperat
         return fromString(name, ModelOperationStatus.class);
     }
 
-    /**
-     * @return known ModelOperationStatus values.
-     */
+    /** @return known ModelOperationStatus values. */
     public static Collection<ModelOperationStatus> values() {
         return values(ModelOperationStatus.class);
     }
-
 }